thesis/requirements.txt
adjustText==1.3.0
asttokens==3.0.0
biopython==1.86
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fire==0.7.1
fonttools==4.59.2
idna==3.11
igraph==0.11.9
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
kiwisolver==1.4.9
line_profiler==5.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
memory-profiler==0.61.0
narwhals==2.7.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.1
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
Pygments==2.19.2
pyintergraph==1.3.3
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
termcolor==3.2.0
texttable==1.7.0
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
UpSetPlot==0.9.0
urllib3==2.5.0
wcwidth==0.2.13
wordcloud==1.9.6
